back member 회원탈퇴시 본인유무권한 확인

+++ src/main/java/com/example/prj1be/controller/MemberController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.SessionAttribute;
 import org.springframework.web.context.request.WebRequest;
 
 @RequestMapping("/api/member")
@@ -75,10 +76,15 @@ public class MemberController {
    }
 
    @DeleteMapping
-   public ResponseEntity delete(String id) {
-      // TODO: 로그인 했는지? -> 안했으면 401에러
-      // TODO: 자기 정보인지? -> 아니면 403 에러
+   public ResponseEntity delete(String id,
+      @SessionAttribute(value = "login", required = false) Member login) {
 
+      if (login == null) {
+         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+      }
+      if (!service.hasAccess(id, login)) {
+         return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+      }
       if (service.deleteMember(id)) {
          return ResponseEntity.ok().build();
       } else {
